mcp-server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
requests==2.32.3
python-dotenv==1.0.1
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
